// src/apispec/src/models/UpdateExposedPortsRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by internal-gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ExposedPortConfig } from './ExposedPortConfig';
import {
    ExposedPortConfigFromJSON,
    ExposedPortConfigFromJSONTyped,
    ExposedPortConfigToJSON,
    ExposedPortConfigToJSONTyped,
} from './ExposedPortConfig';

/**
 * 
 * @export
 * @interface UpdateExposedPortsRequest
 */
export interface UpdateExposedPortsRequest {
    /**
     * 
     * @type {Array<ExposedPortConfig>}
     * @memberof UpdateExposedPortsRequest
     */
    ports: Array<ExposedPortConfig>;
}

/**
 * Check if a given object implements the UpdateExposedPortsRequest interface.
 */
export function instanceOfUpdateExposedPortsRequest(value: object): value is UpdateExposedPortsRequest {
    if (!('ports' in value) || value['ports'] === undefined) return false;
    return true;
}

export function UpdateExposedPortsRequestFromJSON(json: any): UpdateExposedPortsRequest {
    return UpdateExposedPortsRequestFromJSONTyped(json, false);
}

export function UpdateExposedPortsRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateExposedPortsRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'ports': ((json['ports'] as Array<any>).map(ExposedPortConfigFromJSON)),
    };
}

export function UpdateExposedPortsRequestToJSON(json: any): UpdateExposedPortsRequest {
    return UpdateExposedPortsRequestToJSONTyped(json, false);
}

export function UpdateExposedPortsRequestToJSONTyped(value?: UpdateExposedPortsRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'ports': ((value['ports'] as Array<any>).map(ExposedPortConfigToJSON)),
    };
}